# [PyTorch][Vulkan] Reduce 2D float matrix multiplication shader latency by more than 50% on some Android GPUs