Summary:
- Introduce improved algorithm for 2d float GEMM [ output = alpha * (input) * (weight) + beta * (bias) ] that shows more than 50% shader latency reduction on Qualcomm GPUs. Does not apply for the quantized [integer] and batch [3d] matrix multiplication cases.
  - At function call of `run_linear_context()`/`run_addmm_context()`, re-pack the input tensor data to be row-wise element-dense in each texel
  - Reducing global I/O reads and writes through "batching" by fetching 4 input and weight texels each, then performing 16 output computations and writes, in each shader invocation
  - Leverage a loop unrolling/coalescing compile-time optimization of the GLSL->SPIR-V compiler using a macro for 4

Test Plan:
# Numerical Validation
- There are two pre-existing failures on trunk related to conv2d, unrelated to this diff's code paths
- `LD_LIBRARY_PATH=third-party/swiftshader/lib/linux-x64/ buck run fbcode/mode/dev-nosan //xplat/caffe2:pt_vulkan_api_test_bin`

```
[  SKIPPED ] VulkanAPITest.querypool_flushed_shader_log (0 ms)
[----------] 396 tests from VulkanAPITest (38014 ms total)

[----------] Global test environment tear-down
[==========] 396 tests from 1 test suite ran. (38014 ms total)
[  PASSED  ] 393 tests.
[  SKIPPED ] 1 test, listed below:
[  SKIPPED ] VulkanAPITest.querypool_flushed_shader_log
[  FAILED  ] 2 tests, listed below:
[  FAILED  ] VulkanAPITest.conv2d_pw_prepack
[  FAILED  ] VulkanAPITest.conv2d_pw_prepack_bc
```
# Performance Validation with Matrix Multiplication Benchmark Binary
- build the benchmark binary on both this diff and trunk modified for 100 iterations, `buck2 build -c ndk.debug_info_level=0 -c ndk.static_linking=true -c pt.enable_qpl=0 -c pt.vulkan_use_gpu_diagnostics=1 --target-platforms=ovr_config//platform/android:arm32-fbsource //xplat/caffe2:pt_vulkan_mm_perf_test_binAndroid --show-output`


__on local testing against a Samsung Galaxy S22 Ultra **75% reduction**__
- this diff
```
Benchmark                                                                                       Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------
run_linear_context_benchmark/N:500/M:500/P:500/iterations:100/manual_time/threads:1          2.08 ms         10.3 ms          100
```
- trunk:
```
Benchmark                                                                                       Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------
run_linear_context_benchmark/N:500/M:500/P:500/iterations:100/manual_time/threads:1          9.11 ms         13.8 ms          100
```
__on local testing against our Android chipset of interest **50% reduction**__
- this diff:
```
Benchmark                                                                                       Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------
[...]
run_linear_context_benchmark/N:500/M:500/P:500/iterations:100/manual_time/threads:1          40.0 ms         90.6 ms          100

```
- trunk:
```
Benchmark                                                                                       Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------
[...]
run_linear_context_benchmark/N:500/M:500/P:500/iterations:100/manual_time/threads:1          81.3 ms          106 ms          100
```
__on local testing against Google Pixel 7 Pro **55% reduction**__
- this diff:
```
Benchmark                                                                                       Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------
[...]
run_linear_context_benchmark/N:500/M:500/P:500/iterations:100/manual_time/threads:1          7.38 ms         10.7 ms          100
```
- trunk:
```
Benchmark                                                                                       Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------------------------
[...]
run_linear_context_benchmark/N:500/M:500/P:500/iterations:100/manual_time/threads:1          16.2 ms         12.4 ms          100

```

Reviewed By: yipjustin

Differential Revision: D50441864